infra: collect things with wildcard exports
Removes checking for duplicate class names. I think that's meant
to be an error in ES6 modules (i.e. because two exports would be
sharing the same name), but Node.js seemingly is fine with it,
so just be careful for now.
